requirements.txt
altgraph==0.17.2
asgiref==3.4.1
Django==3.2.7
future==0.18.2
pefile==2022.5.30
Pillow==8.3.2
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
pytz==2021.1
pywin32-ctypes==0.2.0
sqlparse==0.4.2
